chromium: adding updated version 24.0.1312.52; refactor build
The chromium build now requires ninja-native to build, as it builds faster than make.
Refactor EXTRA_OEGYP so that it is passed via include.gypi and oe-defaults.gypi files.
This allows for custom per-platform builds. Add support for qemux86 target, as well as
existing armv7a (tested) and armv6 (not tested.)