nss: change shebang line for smime script
This patch fixes the shebang line of the smime script and also adds a
'perl' RDEPENDS for the main package since smime needs perl to run.

Also, this patch fixes some do_rootfs failures when using RPM backend
because RPM automatically scans the files and, if perl scripts are
found, adds dependency on the file found in the shebang line (i.e.
/usr/local/bin/perl).

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@intel.com>