chromium: Disable fatal linker warnings related to SQLite
With the gold linker, problems with hidden sqlite3 symbols occur. Example:

  warning: hidden symbol 'sqlite3_temp_directory' in obj/third_party/sqlite/libsqlite3.a
  (obj/third_party/sqlite/amalgamation/sqlite.sqlite3.o) is referenced by DSO
  /mnt/ramdisk/yocto/qemu-test-build/tmp//sysroots/qemux86/usr/lib/../lib/libsoftokn3.so

The problem is known, and a solution would be to set the build scripts
to not build sqlite3 and use the system's sqlite3 library instead.
However, currently, this is not possible:

  https://code.google.com/p/chromium/issues/detail?id=22208

The Crosswalk project fixed this by setting the warnings to non-fatal:

  https://github.com/crosswalk-project/crosswalk/commit/83d22586852a3d64300fd024e24ade9a4ac325be

The disable_fatal_linker_warnings GYP flag used in that commit is already
part of Chromium 37. It just needs to be enabled. This commit does
precisely that.

Signed-off-by: Carlos Rafael Giani <dv@pseudoterminal.org>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>