chromium: Upgrade to version 38.0.2125.101
* Starting with this version, several cross compilation specific parts in
the GYP files need to be disabled with both additional flags and an extra
patch. Chromium 38 now includes clang specifics, which are enabled by
default. Additional flags (-Dhost_clang=0 -Dclang=0) disable these.
Furthermore, the GYP files now include default values for CC and CXX
(https://codereview.chromium.org/414343003/patch/40001/50001), which
could not be overridden with environment variables. A patch removes this.

* ozone-wayland's Milestone-Harvest branch is the one to use for
Chromium 38. It includes a patch that collides with another one that is
already present in the chromium/ directly. Consequently, this other patch
is now only used for X11 builds.

* Only 00*.patch patch files from ozone-wayland are used, since
the 10*.patch files cannot be applied cleanly and easily (and are not
actually required for builds with this recipe, since another patch already
takes care of the X11 linker flags removal).

Signed-off-by: Carlos Rafael Giani <dv@pseudoterminal.org>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>