chromium: update run script for early run
If the chromium run-helper script (google-chrome) is run as part of a
(sysvinit) init script, shell variables $USER and $HOME will not be set
(leading to unwanted behaviour).

Replace the use of $USER with `whoami` and replace the use of $HOME with the
OE bitbake variable ROOT_HOME.

In this way the chromium run-helper script will work as expected in both
instances (as part of an init script and from the cmdline).

Signed-off-by: Trevor Woerner <twoerner@gmail.com>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>