chromium-wayland: always enable CHROMIUM_ENABLE_WAYLAND flag
It makes no sense to trigger a build of chromium-wayland and not
enable CHROMIUM_ENABLE_WAYLAND, so we are doing it directly in the
recipe now.

Now we know the value of the flag beforehand, we are allowed to
remove some conditionals. Additionally, we check if Wayland is
available before building.

Signed-off-by: Jacobo Aragunde Pérez <jaragunde@igalia.com>