Remove chromium-wayland recipe and related tasks/settings.
This is a first step towards updating the main Chromium recipe to a more
recent milestone. The ozone-wayland project has been unmaintained for a long
time, and we will not be able to keep supporting it as-is without
significant overhead; meanwhile, there has been work upstream to make
Chromium work on Wayland, and we should be able to leverage it within a few
milestones.

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>