chromium: Improve licensing information.
While the Chromium code base itself is BSD-licensed, Chromium is also made
up of several other third-party projects as well as the Blink web engine,
which has a mixture of BSD and LGPL 2.0 and 2.1 licenses.

We now list a lot more license files (including some we may not end up
using), and LICENSE is a mixture of multiple licenses.

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>