chromium: Set $B appropriately and simplify build directory references.
Instead of using ${S}/out/${CHROMIUM_BUILD_TYPE} to refer to the build
directory in the tasks, just set $B appropriately in one place.

By doing so, we ensure we're in that directory when running do_compile() and
do_install(), which allows us to refer to files and directories there
without an absolute path and have a recipe that looks cleaner.

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>