chromium: Start using upstream's wrapper script instead of our own.
The upstream script is more widely tested and maintained, we just need to
replace a few placeholders to get a fully functional wrapper script for
free.

Since the script expects to be placed alongside the actual Chromium binary,
the binary is now installed as chromium-bin and the wrapper is installed as
chromium (with a symlink in ${bindir}).

A small patch has been added for the existing CHROME_EXTRA_FLAGS and
run-as-root functionality to continue working.

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>