chromium: unify recipe and move out upstream Chromium specific logic
The intent of this change is to share as many parts of the recipe as
possible in order to easily switch between different upstream
and downstream versions of Chromium in the future.

To be more specific: chromium-gn.inc and chromium-x11.inc
are introduced.

chromium-gn.inc is used now for all the possible chromium
versions, and chromium-x11.inc will be shared between different
upstream Chromium versions. One just simply makes another
chromium-x11_{VERSION}.bb recipe and everything will be handled
automatically without dramatic changes.

Signed-off-by: Maksim Sisov <msisov@igalia.com>