chromium: introduce chromium-ozone-wayland.
This patch adds a chromium-ozone-wayland recipe to this meta-browser
layer. The recipe is based on the ozone-wayland-dev branch found at
github.com/Igalia/chromium branch, which is being used as a development
branch. The branch is rebased against the Chromium ToT every week and
it's possible that this recipe will have some minor changes every
couple of weeks.

Two additional .inc files are added: chromium-ozone-wayland-tarball.inc
and chromium-ozone-wayland.inc.

The first one contains nothing but licences and an url of the source.
The second has recipe specific dependencies and gn args.

What is more, additional patches are added: a qemu v8 wrapper patch,
which patches chromium side (chromium-gn.inc has already had another
qemu v8 wrapper patch that patches V8, but in the recent ToT both V8
and chromium must be patched), and a patch that rotates gcc and g++
cpp/c extra flags in order to ensure chromium is compiled with right
flags are the original ones are not overriden (this is needed
to compile Chromium in the AGL environment, otherwise some errors
may occur during linking stage).

Signed-off-by: Maksim Sisov <msisov@igalia.com>