chromium-ozone-wayland: Disable Jumbo builds by default.
Whether to use Jumbo builds or not should be up to each user; while it may
decrease build times, it can require more RAM (it tends to cause problems
for me specifically when using GCC because of the amount of memory it
requires).

Additionally, it's not an Ozone-Wayland-specific setting, so it should've
been in chromium-gn.inc had it been the case.

We now provide a PACKAGECONFIG knob (off by default) to control that.

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>