chromium: Stop passing --start-fullscreen with the kiosk-mode PACKAGECONFIG.
The only purpose of the "--start-fullscreen" option is to start Chromium in,
well, full screen mode. It turns out the "--kiosk-mode" does the same
thing and applies some other restrictions ("--start-fullscreen" allows one
to exit full screen mode, "--kiosk-mode" does not, for example).

Since "--start-fullscreen" is essentially a subset of "--kiosk-mode",
passing both on the command line is redundant.

See also:
* https://bugs.chromium.org/p/chromium/issues/detail?id=167430
* https://chromium.googlesource.com/chromium/src/+/64.0.3282.186/chrome/browser/ui/startup/startup_browser_creator_impl.cc#364

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>