chromium-x11: Remove musl-related code.
musl support has not been tested in a very long time, and the patches very
likely no longer apply. It only makes sense to keep musl patches around if
there is work to upstream them, otherwise we end up with an increasing
amount of patches that touch sensitive, low-level areas of the code to
maintain.

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>