chromium: Remove support for GCC 5.
GCC 5 is old by Chromium's standards, and its support for C++14 is not
completely functional.

Keeping the recipe building with GCC 5 requires a lot of effort and
patching (which also affects the build with other compilers), and based on
the discussion in #109 it does not seem like the compiler is being widely
used -- and it helps unblock the upcoming update to Chromium 66, as it was
the last pending item to get everything building with the new milestone.

The recipe is building fine with GCC 6 (the default compiler on Yocto's pyro
branch) and GCC 7 (the default compiler on Yocto's rocko branch).

Test-built with:
* pyro, GCC 6 and MACHINE="raspberrypi3"
* rocko, GCC 7 and MACHINE="intel-corei7-64" and MACHINE="qemuarm64".

Closes #109.

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>