chromium-x11: Fix entangled cross/native build flags
When we use pkgconf instead of pkg-config all sorts of
cflags issues start to appear resulting in compiler errors
which all the due to mixing of target cflags and ldflags
onto native compiler flags.

This patch tries to ensure that target flags are contained
to cross builds alone. This issue is also evident when building
for non-glibc targets e.g. where headers will be different
when compared to build host ( which mostly are glibc based)

Signed-off-by: Khem Raj <raj.khem@gmail.com>