chromium-x11: Update to 69.0.3497.100.
This is the latest M69 stable release at the time of writing, and contains
several security fixes detailed in the M69 release announcements made this
month:
* https://chromereleases.googleblog.com/2018/09/stable-channel-update-for-desktop.html
* https://chromereleases.googleblog.com/2018/09/stable-channel-update-for-desktop_11.html
* https://chromereleases.googleblog.com/2018/09/stable-channel-update-for-desktop_17.html

Test-built with:
* Yocto pyro, GCC 6 and MACHINE="raspberrypi3"
* Yocto rocko, GCC 7 and MACHINE="qemuarm64"
* Yocto sumo, GCC 7 and MACHINE="intel-corei7-64"
chromium-ozone-wayland 71.0.3545.0.r589108 continues to configure fine with
the new gn-native.

Speaking of gn-native, the GN source code was split into its own git
repository upstream, and there are now a few patches to ensure our build
process works (some of which were already upstreamed).

oe-clang-fixes.patch was rebased, with the patch information being pulled
from the version in chromium-ozone-wayland/.

Note that building the recipe for ARM and aarch64 on pyro requires
http://lists.openembedded.org/pipermail/openembedded-core/2018-July/153052.html
to be applied to the gcc recipe.

Moved license files:
* chrome/third_party/chromevox/LICENSE: Moved to
  third_party/chromevox/LICENSE with no alterations.
  <https://chromium-review.googlesource.com/1108553>

Updated license files:
* third_party/ffmpeg/CREDITS.chromium: The order of the copyright notices
  was changed.
  <https://chromium.googlesource.com/chromium/third_party/ffmpeg/+/2007874ddb778b95f05317319b511d0bd279b296>

* third_party/jinja2/LICENSE: Line breaks converted from CRLF to LF.
  <https://chromium-review.googlesource.com/993812>

Removed license files:
* third_party/shaderc/LICENSE: No longer user or shipped.
  <https://chromium-review.googlesource.com/1100298>

* third_party/vulkan/LICENSE: Vulkan was updated, and the directory only
  contains headers licensed under the Apache 2.0 license.
  <https://chromium-review.googlesource.com/c/chromium/src/+/1090571>

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>