chromium-ozone-wayland: Stop depending on the system's wayland-scanner binary
So far, we were using Chromium's default GN path for the system
wayland-scanner binary, "/usr/bin/wayland-scanner", which may not be present
at all and which should never be used anyway. Instead, we now point to the
version built by the wayland-native recipe.

Doing so of course requires depending on wayland-native in the first place,
and that requires adding back REQUIRED_DISTRO_FEATURES and a DEPENDS block
that got removed with no explanation when the recipe was updated to
M72 (commit 97876fab, "chromium: Update to 72.0.3626.81").

Fixes #241

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>