chromium: Update to 79.0.3945.79
This update includes the following chromium release:
https://chromereleases.googleblog.com/2019/12/stable-channel-update-for-desktop.html

Chromium Ozone/Wayland:
----------------------
We have removed V4L2 support as we do not test it and it always breaks.
Supporting V4L2 requires time for rebasing and testing. Given that the patch is
not upstreamable and V4L2 implementation always receives a bunch of changes, it
was decided to remove the patches from the recipe.

Test-built with:
  * chromium-x11
    * thud, MACHINE="intel-corei7-64"
    * thud, MACHINE="intel-core2-32"
    * warrior, MACHINE="intel-corei7-64"
    * warrior, MACHINE="intel-corei7-64"
  * chromium-ozone-wayland:
    * thud, MACHINE="qemuarm64"
    * thud, MACHINE="m3ulcb"
    * warrior, MACHINE="raspberrypi3"

License changes
---------------
Removed licenses:
* third_party/expat/files/COPYING:
  <https://crrev.com/c/1822703>

New licenses:
* third_party/expat/src/files/COPYING:
  <https://crrev.com/c/1822703>

Fixes #339

Signed-off-by: Maksim Sisov <msisov@igalia.com>