chromium: Replace some custom patches with upstream backports.
stdint.patch and std_size_t.patch were never upstreamed, and commit
80205badc461 ("chromium: update to 83.0.4103.61") already ended up making
some needless changes to them.

Replace them with backports so that it is easier to know when those patches
can be removed altogether (in some cases, the changes were already part of
the Chromium version we are packaging).

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>