chromium: Add a unified ozone wayland/x11/headless recipe for Chromium
Renamed chromium-ozone-wayland as chromium-ozone and enabled building
x11 and/or wayland and/or headless withing a single recipe.

Depending on DISTRO_FEATURES, chromium-ozone will build chromium source
with x11, wayland or headless backends. If both x11 and wayland is
enabled, both will be supported with single recipe. If none is
available, only headless will be built.