chromium: add patch that enables VA-API with ozone/wayland.
Not only the HWs were tested. The patch was mostly tested
with Intel Gen8 and Gen9. Also, with AMD RX460. So, bear that
in mind when you enable va-api.

In order to have it running, one must enable VA-API and
proprietary codecs in Chromium
PACKAGECONFIG_pn-chromium-ozone-wayland_append = " proprietary-codecs
use_vaapi".