chromium: Backport patch to stop including <sys/random.h>
This header might not exist on systems with very old glibc packages (e.g.
CentOS 7) when building some host binaries that we need to run as part of
the build. It turns out the header is not necessary at all, as Chromium uses
LSS (Linux Syscall Support) to call the syscall and falls back to reading
from /dev/urandom when that fails.

Fixes #445.

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>