WIP: chromium: move dbus to PACKAGECONFIG
dbus is a (r)dpendency of chromium as it tries to
communicate with different dbus services on launch.

When running it with a popular DE such as xfce
no issue occurs.

However, when trying to run a minimal embedded build,
(only Xserver, no DE & kiosk-mode as a service), chromium complains
about missing `upower` and missing environment variables such as
DBUS_SESSION_BUS_ADDRESS

This patch attempts to make the dependency to dbus optional
by reapplying a patch on chromium-85
from https://github.com/smaeul/portage-overlay/
    blob/master/www-client/chromium/files/chromium-85-optional-dbus.patch

Signed-off-by: Yi Fan Yu <yifan.yu@windriver.com>