chromium: Remove impl-side-painting PACKAGECONFIG knob.
This knob was added back in 2014 in commit fb6df963376c ("chromium: Add
recipe for version 37.0.2062.0"). Multithreaded rasterization was under
development back then, but this hasn't been the case since M39 and M40 (see
https://crbug.com/169282 and https://crbug.com/413479).

Given multithreaded rasterization was enabled by default in 2015, the
--enable-impl-side-painting was removed around that time and that this knob
was off by default, it should be safe to remove it altogether.

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>