chromium: add PACKAGECONFIG for gtk
Recently i found out, that chromium needs to be built with use_gtk=true if you want it to respect the gtk cursor theme.
This seems to be related to a change in use flags in chromium. There has been a use_gtk3 option that was true by
default. In addition we had use_gtk to enable builds with gtk+2.

Now it looks like we only have use_gtk to enable build with gtk+
https://github.com/chromium/chromium/blob/72ceeed2ebcd505b8d8205ed7354e862b871995e/build/config/linux/gtk/gtk.gni#L9

We have this disabled by default so i guess all current builds are without using gtk:
https://github.com/OSSystems/meta-browser/blob/8df26e66c28b43f1630fcd2cb952a1e69b6e6558/meta-chromium/recipes-browser/chromium/chromium-ozone-wayland_94.0.4606.71.bb#L30

Make gtk a PACKAGECONFIG to have an option.

Signed-off-by: MarkusVolk <f_l_k@t-online.de>