chromium-ozone-wayland: Fixed terminate caused by binding to wrong version.
The Ozone/Wayland implementation had a few places where the Wayland
objects were bound without proper checking for their versions.  That was
part of the technical debt not addressed before, and ended up causing
the issue explained in the linked crbug: the compositor terminates the
client that binds to the protocol that it does not actually support.

This patch fixes the issue by adding the necessary checks in all places
where they were missing.  Also a convenience macro for validating the
version is proposed.

author	Alexander Dunaev

https://bugs.chromium.org/p/chromium/issues/detail?id=1279574

Signed-off-by: Markus Volk <f_l_k@t-online.de>