chromium: Add a generate_license_list.py script to help generate LIC_FILES_CHKSUM.
This is based on an original version I wrote years ago. Specifically, its
`SKIPPED_DIRECTORIES` variable might need to be updated.

As described in the script, it uses Chromium's own tools/licenses.py script
to scan for third_party directories and license files. This means its output
is generated on a best-effort basis, as some directories are non-compliant
upstream or may not be found. It might also include directories which are
not used in a Yocto/OpenEmbedded build.