chromium-x11: Add RDEPENDS on libx11-xcb.
libX11-xcb.so.1 is loaded at runtime by the X11 backend, but we had no
explicit dependency on the package that provides it (libx11-xcb).

It ended up being installed most of the time anyway, but make sure it is
listed as a runtime dependency to avoid the issue described in #683.

Fixes #683.

Signed-off-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>