chromium: Drop patches that are specific to clang < 17
Given that we have decided not to support kirkstone until it uses a more
recent version of clang, we can now drop all patches that are only
needed to fix errors with clang < 17 (as nanbield uses clang 17).

This also updates LAYERSERIES_COMPAT, dropping kirkstone, as well as the
EOL'd mickledore and langdale.

Signed-off-by: Max Ihlenfeldt <max@igalia.com>