chromium: Merge interlinked compiler option patches
Build and patch changes:
------------------------

`arm/0001-BUILD-do-not-specify-march-on-arm.patch` depends on
`0001-Remove-the-GN-settings-done-for-clang-that-conflict-.patch` (i.e.
it fails to apply on its own), which is confusing and makes rebasing on
new upstream releases harder.

This merges the two patches into a single one.

License changes:
----------------

Added licenses: none.

Removed licenses: none.

Updated licenses: none.

Test-built:
-----------

* chromium-ozone-wayland:
 - scarthgap, clang, MACHINE=qemuarm64

* chromium-x11:
 - master, clang,    MACHINE=qemuarm

Signed-off-by: Max Ihlenfeldt <max@igalia.com>