firefox: move build related config to separate repository
Move test and build related configs to a separate layer/repo (meta-firefox-test)

It started to pollute this layer, and although it is important,
the test and build related recipes and their settings started to
become too opinionated for this layer, and they also started to
overflow to other recipes.

Keep them separate, so they won't break other things accidentally.