firefox: update 135.0 -> 135.0.1, improve build
- move rust_gen_targets before configure

This seems to be causing the problem, that running do_configure
on its own fails: rust_gen_targets was running before do_compile,
but not necessarily before do_configure. It seems that when all
tasks were executed in one go, than this task is executed before
do_configure, so it works. But when do_configure is executed on
its own, then this task is scheduled to be executed after
do_configure, which makes do_configure fail.

To avoid it, reschedule it to be executed always before do_configure.

- remove panic strategy from Cargo.toml and from readme

Removing the panic strategy from the Cargo.toml file should
should be enough to compile Firefox successfully (instead of compiling
Rust with this panic strategy, which is not default in Yocto)