firefox: update 136.0.3 -> 137.0
Firefox 137 requires at least Rust 1.82. Due to the old Rust provided
by meta-lts-mixins layers, meta-rust is needed to compile the latest
Firefox.

This requires some small changes:
- In mozilla.bbclass the value of RUST_TARGET and RUST_HOST are set with
weak assignement, so they can be easily overriden from other configs.
The reason for this is the different target-triple schemes used by
meta-lts-mixins (and oe-core) and meta-rust. The default value is set to work with
meta-lts-mixins (and oe-core), but it is easy to change it to work
with meta-rust.
- To avoid unnecessary Rust pollution on the build system, the mandatory
layer dependency on meta-lts-mixins have been dropped, at least
temporarily.
- CRATE_CC_NO_DEFAULTS is set in the mozilla.bbclass. This is used by
default by meta-lts-mixins and oe-core, but not by meta-rust. Without
this Firefox is compiled by host-default compiler flags, which fails
cross-compilation.