firefox: update 128.10.1esr -> 128.11.0esr
This commit also removes Styhead compatibility.
Styhead is EOL, and as a bigger technical issue, the Rust version available
for that release is just too old for new Firefox (though ESR can be still
compiled, the latest version of Firefox has been inoperable since a couple
of releases). It is time to remove it.

Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>