chromium: Add a chromium build/test github action
Added a workflow for Chromium compile/test which is inspired from the
Firefox version. I've restricted the build matrix to avoid building
targets which we are not interested in. But meta-chromium-test provides
recipes for building all possible combinations if neeeded.

meta-chromium-test repo has the necessary infrastructure for running
a basic smoke test. But test is not sufficient enough, so I decided to
not to run the test step.

Removed -g flags from arm builds. arm linker is unable to link the image
if -g flag is used.