chromium: 140.0.7339.136
RUNTIME and TC_CXX_RUNTIME are both set explicitly to "llvm"
in chromium.inc. Therefore the :runtime-llvm override is
not needed because it's always on. But it's also a problem in
whinlatter now, because the DEPENDS = "compiler-rt" is not
applied properly, making do_copy_clang_library() fail.
Remove :runtime-llvm override.

Use TOOLCHAIN_NATIVE instead of TOOLCHAIN:class-native.

Drop 0008-Use-the-correct-path-to-libclang_rt.builtins.a.patch
and fix do_copy_clang_library() in a way that's compatible with
clang in oe-core.

Depend on lld-native.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>