firefox: remove crate includes
Initially, when the crates include was added, Firefox's vendored crates
were not enough to build Firefox, it required downloading a number of
other crates to be able to compile without failure.

Now however it looks that the tarball contains all required crates, so
use it instead of vendoring them from the recipe.

Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>