firefox/wasi-sdk: upgrade wasi-sdk 21->29
This update is an optional build-time requirement for Firefox 144+ to be
able to build some sandboxed libraries.

Since the build system was overhauled by upstream, the original recipe
was split into two also: one recipe builds the toolchain (which is essentially
a WebAssembly-only clang), and the second recipe populates the actual sysroot.

Though they use the same main CMakeLists.txt file, the environment is quite
different for the two steps, hence the split.

One notable change is that building this recipe requires at least Clang 18,
which is a bump from the previous Clang 14. Though Kirkstone can still build
it, it requires switching to a new meta-clang branch.

Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>